requirements.txt
cachelib==0.10.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Session==0.4.0
gunicorn==20.1.0
idna==2.8
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
JsonDecoder==0.0.8
jsonpickle==3.0.1
MarkupSafe==2.1.1
oauthlib==3.0.1
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.6
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
requests==2.21.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.24.3
uuid0==0.2.7
Werkzeug==2.2.2
pytest==7.4.3
